+++ package.json
@@ -18,9 +18,10 @@
     "@expo/vector-icons": "^10.0.6",
     "@react-native-community/masked-view": "^0.1.7",
     "@react-navigation/bottom-tabs": "^5.2.4",
+    "@react-navigation/core": "^5.3.0",
+    "@react-navigation/drawer": "^5.3.4",
     "@react-navigation/native": "^5.1.3",
     "@react-navigation/stack": "^5.2.8",
-    "@react-navigation/core": "^5.3.0",
     "expo": "^36.0.2",
     "expo-asset": "^8.1.2",
     "expo-constants": "^9.0.0",
@@ -36,7 +37,6 @@
     "react-native-web": "^0.12.2"
   },
   "devDependencies": {
-    "react-test-renderer": "~16.11.0",
     "@babel/core": "^7.9.0",
     "@types/jest": "^25.1.4",
     "@types/react": "^16.9.27",
@@ -45,6 +45,7 @@
     "babel-preset-expo": "^8.1.0",
     "jest": "^25.2.3",
     "react-navigation": "^4.3.5",
+    "react-test-renderer": "~16.11.0",
     "ts-jest": "^25.2.1",
     "typescript": "^3.8.3"
   },
