+++ screens/homeScreen/inventoryScreen/InventoryScreen.tsx
@@ -1,0 +1,10 @@
+import * as React from "react";
+import {Text, View} from "react-native";
+
+export const Inventory = () => {
+    return (
+        <View>
+            <Text>Сюда надо Запихнуть инвентарь</Text>
+        </View>
+    )
+}
